invalidate pooled connections after 404
we suspect there may have been a problem where
a keep alive connection was made then a DNS change
occurred but we kept sending requests to an old
server that started to return 404s

since we do not have a good way of invaliding http
connections from outside of lhttpc I have decided
to hardcode the logic to junk connections after
receiving a 404.

i think the only other reasonable alternative
would be to expose an interface to remove all
connections from the pool matching a hostname
and then the client could decide whether the
connection needs invalidation based on its own
suspicions. 

in a perfect world the client could decide
whether to invalidate on a connection by
connection basis but this doesn't seem possible
because the lhttpc interface is a request/response
interface and doesn't expose connection handling
to the client. however, one way of doing it would
be to allow the client to specify a callback which
would allow the client to determine whether the 
connection is 'bad' or not before returning the
connection to the pool.